avance de sql20
debo arreglar orden pq no encuentra id_cliente

+++ ecommerce-backend/src/main/java/tbd/lab1/entities/DetalleOrdenEntity.java
@@ -7,7 +7,7 @@ import java.math.BigDecimal;
 
 public class DetalleOrdenEntity {
 
-    private Long idDetalle;
+    private Integer idDetalle;
 
     private OrdenEntity orden;
